Hacer el cambio de edad a fecha de nacimiento en el modulo de Autores

+++ grails-app/controllers/Colecciones/AutorController.groovy
@@ -3,6 +3,8 @@ package Colecciones
 import grails.plugin.springsecurity.annotation.Secured
 import grails.transaction.Transactional
 
+import java.text.SimpleDateFormat
+
 @Transactional(readOnly = true)
 class AutorController {
 
@@ -80,7 +82,6 @@ class AutorController {
     @Transactional
     def save(Autor autorInstance, params) {
 		def validadorForm = autorService.validarForm(params)
-		def validadorFoto = null
 		//Validar los datos del formulario
         if (validadorForm.error) {
 			flash.error = message(code: validadorForm.mensaje)
@@ -100,7 +101,8 @@ class AutorController {
 			redirect(action: "create")
 			return
 		}
-		//Si hay foto guardarla en la carpeta configurada 
+		//Si hay foto guardarla en la carpeta configurada
+		def validadorFoto = null
 		def file = request.getFile('imagen')
 		if(!file.empty){
 			validadorFoto = coleccionesService.saveImg(file, params.nombre, params.apellido, false)
@@ -117,13 +119,20 @@ class AutorController {
 
 			autorInstance.fechaInscripcion = new Date()
 			autorInstance.ultimaModificacion = new Date()
+			autorInstance.fechaNacimento = new SimpleDateFormat("dd/MM/yy").parse(params.fechaDeNacimento)
 			autorInstance.difunto = (params.difunto.equals("true")?true:false)
 			if(autorInstance.save(flush:true)){
 				log.info "Creando entrada en el historial de un nuevo autor"
 				flash.message = message(code: "autores.message.save.ok", args: [nombreAutor])
 				registerHistorialService.registrarAutor(autorInstance, 0)
 			}else{
-				coleccionesService.deleteImage(validadorFoto.path)
+				if(!file.empty){
+					coleccionesService.deleteImage(validadorFoto.path)
+				}
+				log.error "No se ha podido guardar en base de datos el Autor"
+				flash.error = message(code: "autores.errores.save.bbdd")
+				redirect(action: "create")
+				return
 			}
 		}catch(Exception e){
 			log.error "No se ha podido guardar en base de datos " + e
@@ -212,11 +221,17 @@ class AutorController {
 			def nombreAutor = params.nombre + " " + params.apellido
 
 			autorInstance.ultimaModificacion = new Date()
+			autorInstance.fechaNacimento = new SimpleDateFormat("dd/MM/yy").parse(params.fechaDeNacimento)
 			autorInstance.properties = params //Guardamos los datos restantes
 			if(autorInstance.save(flush:true)){
 				log.info "Creando entrada en el historial de una actualizacion de un autor"
 				flash.message = message(code: "autores.message.update.ok", args: [nombreAutor])
 				registerHistorialService.registrarAutor(autorInstance, 1)
+			}else{
+				log.error "No se ha podido guardar en base de datos el Autor"
+				flash.error = message(code: "autores.errores.update.bbdd")
+				redirect(action: "edit", id:autorInstance.id)
+				return
 			}
 		}catch(Exception e){
 			log.error "No se ha podido guardar en base de datos " + e
